+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 - ✨(back) add model fallback mechanism
 - ✨(back) add celery for running background tasks
 - 🧱(helm) add celery worker and beat deployments
+- ✨(front) add "Edit in Docs" action to export a message to Docs
+
 
 ### Changed